Refactored the public interface.
And all code that depended on it (mostly tests and lesson classes).
Started working on abstractexercise, but not gotten very far -- it's not
a very promising line of work.

Lessons still need to be updated, and should be done ASAP.  A build
system also needs to be put in place, testing lessons is too troublesome
as things are.